+++ packages/keyring-api/src/api/keyring.ts
@@ -8,6 +8,8 @@ import type { KeyringAccount } from './account';
 import type { ResolvedAccountAddress } from './address';
 import type { Balance } from './balance';
 import type { CaipChainId, CaipAssetType, CaipAssetTypeOrId } from './caip';
+import type { DiscoveredAccount } from './discovery';
+import type { EntropySourceId } from './entropy';
 import type { KeyringAccountData } from './export';
 import type { MetaMaskOptions } from './options';
 import type { Paginated, Pagination } from './pagination';
@@ -83,6 +85,30 @@ export type Keyring = {
     id: string,
     pagination: Pagination,
   ): Promise<Paginated<Transaction>>;
+
+  /**
+   * Discover accounts.
+   *
+   * This method is called by the client to allow the keyring to discover
+   * existing accounts based on the provided scopes and entropy source ID. Are
+   * considered existing accounts, accounts that have at least one transaction,
+   * as per BIP-44.
+   *
+   * The `groupIndex` is used to group accounts with the same value. In
+   * strictly BIP-44 wallets, it matches `account_index`, but in wallets that
+   * deviate from BIP-44 recommendations, it may align with a different path
+   * level for compatibility.
+   *
+   * @param scopes - The list of scopes for account discovery.
+   * @param entropySource - The ID of the entropy source used to derive the accounts.
+   * @param groupIndex - The group index that should be used to derive the accounts.
+   * @returns A promise resolving to a list of discovered accounts.
+   */
+  discoverAccounts?(
+    scopes: CaipChainId[],
+    entropySource: EntropySourceId,
+    groupIndex: number,
+  ): Promise<DiscoveredAccount[]>;
 
   /**
    * Retrieve the balances of a given account.